requirements.txt
APScheduler==3.6.3
asgiref==3.8.1
cachetools==4.2.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.20
djangorestframework==3.16.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==24.2
pillow==11.2.1
psycopg2-binary==2.9.10
python-telegram-bot==13.7
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tornado==6.4.2
typing_extensions==4.13.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
